계산기 기능 추가, pruning 오류 수정, VoiceConnection 에러 이벤트 대응

+++ commands/song_pruning.js
@@ -4,6 +4,10 @@ module.exports = {
     description: "- 봇의 음악 메시지 자동정리 기능 상태를 전환합니다.",
     type: ["음악"],
     async execute(message) {
+        if (!message.guild) {
+            return message.reply("사용이 불가능한 채널입니다."); // 그룹톡 여부 체크
+        }
+
         const find = await db.get("SELECT * FROM pruningskip WHERE channelid = ?", [message.guild.id]);
         if (find) { // 기존상태: OFF
             await db.run("DELETE FROM pruningskip WHERE channelid = ?", [message.guild.id]);
